Move `eval_f` to arith
